+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### ENHANCEMENTS:
 
+- feat(service): improve Fastly 404 error message for service permissions [#964](https://github.com/fastly/terraform-provider-fastly/pull/964)
 - feat(logging): Restore support for 'placement' attribute
 
 ### BUG FIXES: